Add bonjour detail browser

+++ EMW ToolBox/CustomIOS7AlertView/CustomIOS7AlertView.h
@@ -41,7 +41,7 @@
 - (void)close;
 
 - (IBAction)customIOS7dialogButtonTouchUpInside:(id)sender;
-- (void)setOnButtonTouchUpInside:(void (^)(CustomIOS7AlertView *alertView, int buttonIndex))onButtonTouchUpInside;
+- (void)setOnButtonTouchUpInside:(void (^)(CustomIOS7AlertView *alertView, NSInteger buttonIndex))onButtonTouchUpInside;
 
 - (void)deviceOrientationDidChange: (NSNotification *)notification;
 - (void)dealloc;
