+++ EMW ToolBox/CustomIOS7AlertView/CustomIOS7AlertView.m
@@ -132,7 +132,7 @@ CGFloat buttonSpacerHeight = 0;
 // Default button behaviour
 - (void)customIOS7dialogButtonTouchUpInside: (CustomIOS7AlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex
 {
-    NSLog(@"Button Clicked! %d, %ld", buttonIndex, (long)[alertView tag]);
+    NSLog(@"Button Clicked! %ld, %ld", (long)buttonIndex, (long)[alertView tag]);
     [self close];
 }
